+++ lib/src/gsmtap/mod.rs
@@ -3,6 +3,8 @@
 use deku::prelude::*;
 use num_enum::TryFromPrimitive;
 
+pub mod parser;
+
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub enum GsmtapType {
     Um(UmSubtype),
